Move the ffmpeg interface to ffmpeg.py
